requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# LangGraph and LangChain for multi-agent system
langgraph>=0.0.55
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# Google Gemini API
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0

# Search capabilities
duckduckgo-search>=4.0.0

# Data processing and utilities
pydantic>=2.0.0
typing-extensions>=4.8.0
asyncio-throttle>=1.0.0

# Optional: Enhanced functionality
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0